requirements-docker.txt
# Core FastAPI dependencies
fastapi==0.109.0
starlette==0.35.1
uvicorn==0.27.0
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.9

# FastAPI/Starlette transitive dependencies
annotated-types==0.7.0
anyio==4.11.0
sniffio==1.3.1
httptools==0.7.1
h11==0.16.0
watchfiles==1.1.1
websockets==15.0.1
typing_extensions==4.15.0

# PyTorch and ML dependencies
torch==2.9.1
torchvision==0.24.1
numpy==2.2.6
ml_dtypes==0.5.3

# Computer Vision
opencv-python==4.12.0.88
pillow==11.3.0

# Face detection
uniface==1.1.0

# Common dependencies
packaging==25.0
setuptools==70.2.0
filelock==3.19.1

# Webhook and environment variable support
httpx==0.27.0
python-dotenv==1.0.0